Fix inductor `end` of slice.
This PR fixes checks for the `end` integer when slicing a tensor. The problem was that
slicing dimension `dim` with `(start, end, step)` created (among others) the guard: `end <
size[dim]`.

Although, that doesn't generate incorrect results, a tighter guard would be: `end <
size[dim] + 1`, since `end` is not included in the slice.

ghstack-source-id: a3a5d2139a68d362d2e076a26e78b28a7ab1ea50
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106208